fix: clean up codebase
