+++ src/compiler/parser/mod.rs
@@ -2,9 +2,11 @@ mod syntax_module;
 mod pattern;
 mod preset;
 mod util;
+mod metadata;
 
 pub use syntax_module::*;
 pub use pattern::*;
+pub use metadata::*;
 pub mod patterns {
     pub use super::pattern::*;
     pub use super::preset::*;
