+++ tests/arith_modules/mod.rs
@@ -1,0 +1,6 @@
+mod expr;
+mod number;
+mod add;
+pub use expr::*;
+pub use number::*;
+pub use add::*;
